+++ Options/OptAddr.h
@@ -27,7 +27,7 @@ class TOptAddr : public TOpt
  public:
     TOptAddr(int type, const char * buf, unsigned short len, TMsg* parent);
     TOptAddr(int type, SPtr<TIPv6Addr> addr, TMsg * parent);
-    int getSize();
+    size_t getSize();
     char * storeSelf( char* buf);
     SPtr<TIPv6Addr> getAddr();
     virtual bool doDuties() { return true; } // does nothing on its own
